[LOG4J2-1699] Refactor PosixFileAttributeViewAction with Builder rather than method factory. Split posix tests inner static classes to 2 differents test class file because it doesnt run on jenkins. Add a static method in FileUtils to check if posix file attribute view is supported. It will allow later to log if posix attributes are defined but the file system doesnt support them. Fix when user has only one group in the test case.
